+++ CHANGES.rst
@@ -10,7 +10,10 @@ Breaking changes:
 
 New features:
 
-- *add item here*
+- pat-resourceregistry:
+  - Add input fiels for ``merge_with``, ``last_compilation``, ``jscompilation`` and ``csscompilation``.
+  - Add ``load_async`` and ``load_defer`` options.
+  [thet]
 
 Bug fixes: